[Refactor] Divide Package
